requirements.txt
pbr
transmissionrpc>=0.8
python-daemon
